Day5: Funcatinality ready, CSS changes added

+++ src/components/Comment.jsx
@@ -11,6 +11,14 @@ class Comment extends Component {
     // }
 
     clickDelete = () => {
+        if (!this.props.user) {
+            window.alert('You are not logged in');
+            return;
+        }
+        if (this.props.user !== this.props.author) {
+            window.alert('Not permitted. You can delete your comments only');
+            return;
+        }
         const { comment_id, deleteComment } = this.props;
         deleteComment(comment_id)
     }
@@ -18,14 +26,13 @@ class Comment extends Component {
     render() {
         const { comment_id, body, author, created_at, votes, deleteComment } = this.props;
         console.log('from comment', comment_id)
-        return (<li key={comment_id}>
-            <p>{body}</p>
-            <p>{author}</p>
-            <p>{created_at}</p>
+        return (<li className='comment-grid-container' key={comment_id}>
+            <p className='comment-body' >{body}</p>
+            <p className='comment-header' >Posted by: {author} at {created_at.split('T')[0]}</p>
             {/* <p>{votes}</p> */}
-            <Voter votes={votes} id={comment_id} type='comments' />
+            <Voter votes={votes} id={comment_id} type='comments' user={this.props.user} />
 
-            <Avatar>
+            <Avatar className='delete-comment'>
                 <Delete onClick={this.clickDelete} />
             </Avatar>
             {/* <Link to='/articles'></Link> */}
