+++ src/components/SingleArticleView.jsx
@@ -6,13 +6,20 @@ import { Delete } from '@material-ui/icons';
 import Voter from './Voter';
 import Commentlist from './CommentList';
 import AddComment from './AddComment';
+import Button from '@material-ui/core/Button';
 
 
 
 class SingleArticleView extends Component {
-    state = {
-        commentsOpened: false,
-        article: null
+    constructor() {
+        super();
+        this.state = {
+            commentsOpened: false,
+            article: null,
+            comments: []
+        }
+        this.getComments = this.getComments.bind(this);
+        this.deleteArticle = this.deleteArticle.bind(this);
     }
 
 
@@ -21,19 +28,16 @@ class SingleArticleView extends Component {
         axios.get(url).then(({ data: { article } }) => this.setState({ article: article }))
     }
 
+    getComments(id) {
+        const url = `https://dry-island-66406.herokuapp.com/api/articles/${id}/comments`;
+        axios.get(url).then(({ data: { comments } }) => this.setState({ comments: comments }))
+    }
+
     componentDidMount() {
         const { id } = this.props;
         this.getArticle(id);
     }
 
-    componentDidUpdate(prevProps, prevState) {
-        const { id } = this.props;
-        if (prevProps.id !== id) {
-            this.getArticle(id);
-            this.getComments(id);
-        }
-
-    }
     openComments() {
         this.setState({
             commentsOpened: true
@@ -46,12 +50,22 @@ class SingleArticleView extends Component {
     }
 
     deleteArticle() {
+        console.log(this.props);
+        if (!this.props.user) {
+            window.alert('You are not logged in');
+            return;
+        }
+        if (this.props.user !== this.state.article.author) {
+            window.alert('Not permitted. You can delete only your articles');
+            return;
+        }
         const { id } = this.props;
         const url = `https://dry-island-66406.herokuapp.com/api/articles/${id}`;
         axios.delete(url)
             .then((response) => {
                 if (response.status === 200)
                     console.log('successfully deleted')
+                this.props.getArticles();
             })
     }
 
@@ -74,14 +88,17 @@ class SingleArticleView extends Component {
                 <p>{this.state.article.topic}</p>
                 <p>{this.state.article.body}</p>
                 <Voter votes={this.state.article.votes} id={this.props.id} type='articles' />
-                <p>Comments: {this.state.article.comment_count}</p>
+                <p>Comments: {this.state.comments.length || this.state.article.comment_count}</p>
                 <p>Created at: {this.state.article.created_at}</p>
-                {!this.state.commentsOpened ? <button onClick={() => this.openComments()}>Open comments</button> : <button onClick={() => this.closeComments()}>Close comments</button>}
-                {!this.state.commentsOpened ? <div></div> : <Commentlist id={this.props.id} />}
+                {!this.state.commentsOpened ?
+                    <Button variant="outlined" onClick={() => this.openComments()}>OPEN COMMENTS</Button> :
+                    <Button variant="outlined" onClick={() => this.closeComments()}>CLOSE COMMENTS</Button>}
+                {!this.state.commentsOpened ? <div></div> : <>
+                    <AddComment id={this.props.id} getComments={this.getComments} comments={this.state.comments} user={this.props.user} />
+                    <Commentlist id={this.props.id} getComments={this.getComments} comments={this.state.comments} user={this.props.user} /> </>}
                 <Link to='/articles'><Avatar>
-                    <Delete onClick={() => this.deleteArticle()} />
+                    <Delete onClick={() => this.deleteArticle()} className='delete' />
                 </Avatar></Link>
-                <AddComment id={this.props.id} />
             </div></div>)
     }
